requirements.txt
Django>=5.2
psycopg2-binary
django-bootstrap5
Pillow>=10.0.0
numpy
pandas
prophet
celery
redis
gunicorn>=22.0
whitenoise>=6.7
dj-database-url>=2.1
python-dotenv>=1.0
polib>=1.2.0
